docs(c-memory-lab): 修正技术细节与平台描述
将 Mac 架构从 x86-64 更正为 ARM64 (Apple Silicon)，修正 printf 类型提升导致的未定义行为，补充严格别名、有效类型、补码、指针合法性等 C 标准精确描述，避免学生形成错误认知。

+++ c-memory-lab/lessons/01_bytes.c
@@ -5,7 +5,7 @@ static void print_bytes(const void *address, size_t size)
     const unsigned char *bytes = address;
 
     for (size_t i = 0; i < size; ++i) {
-        printf("%02x ", bytes[i]);
+        printf("%02x ", (unsigned int)bytes[i]);
     }
     putchar('\n');
 }
@@ -18,7 +18,7 @@ int main(void)
 
     printf("对象       地址            大小   值\n");
     printf("letter     %p   %zu     %c\n", (void *)&letter, sizeof letter, letter);
-    printf("number     %p   %zu     %#x\n", (void *)&number, sizeof number, number);
+    printf("number     %p   %zu     %#x\n", (void *)&number, sizeof number, (unsigned int)number);
     printf("decimal    %p   %zu     %.2f\n\n", (void *)&decimal, sizeof decimal, decimal);
 
     printf("number 的内存字节（从低地址到高地址）：\n");
